Changed to view_survey_results.php, made year drop down dynamic so it displays 2000 to the current year and real static classes added for the mean time, currently working on pulling real classes with php. Still need to make class changes to google form.

+++ views/curricula/curricula.php
@@ -1,12 +1,25 @@
 <?php
-authorizedPage();
-requireRole(Role::Admin | Role::SuperAdmin);
+/**
+ * PEP Capping 2017 Algozzine's Class
+ *
+ * Page that displays the current curricula.
+ *
+ * This page displays curricula in a grid-like view
+ * where users can then view, edit, or delete
+ * curricula that appear. Superusers can view archived
+ * curricula and restore them or fully delete them.
+ *
+ * @author Jack Grzechowiak
+ * @copyright 2017 Marist College
+ * @version 0.6
+ * @since 0.1
+ */
 
 global $params, $route, $view;
 
 include ('../models/Notification.php');
 
-$pages = ['view','edit','create','delete','restore'];
+$pages = ['view','edit','create','delete','archive', 'classes'];
 
 # Update page title to reflect route
 if (!empty($params) && in_array($params[0], $pages)) {
@@ -19,10 +32,14 @@ if (!empty($params) && $params[0] == 'view') {
     $view->display('curricula/curricula_view.php');
 } else if (!empty($params) && $params[0] == 'edit') {
     $view->display('curricula/curricula_modify.php');
+} else if (!empty($params) && $params[0] == 'classes') {
+    $view->display('curricula/curricula_add_class.php');
 } else if (!empty($params) && $params[0] == 'create') {
     $view->display('curricula/curricula_modify.php');
 } else if (!empty($params) && $params[0] == 'delete') {
-    $view->display('curricula/curricula_archive.php');
+    $view->display('curricula/curricula_delete.php');
+} else if (!empty($params) && $params[0] == 'archive') {
+    $view->display('curricula/curricula_archives.php');
 } else {
     include('header.php');
     global $db;
@@ -30,28 +47,35 @@ if (!empty($params) && $params[0] == 'view') {
     $filter = "";
     if ($_SERVER['REQUEST_METHOD'] == 'POST') {
         $filter = isset($_POST['filter']) ? "%" . $_POST['filter'] . "%" : "%%";
-        $result = $db->query("SELECT * FROM curricula WHERE LOWER(curriculumname::text) LIKE LOWER($1) " .
+        $result = $db->query("SELECT * FROM curricula WHERE df = 0 AND LOWER(curriculumname::text) LIKE LOWER($1) " .
             "OR LOWER(curriculumtype::text) LIKE LOWER($1) ORDER BY curriculumname", [$filter]);
     } else {
-        $result = $db->query("SELECT * FROM curricula ORDER BY curriculumname", []);
+        $result = $db->query("SELECT * FROM curricula WHERE df = 0 ORDER BY curriculumname", []);
     }
 
     ?>
     <div style="width: 100%">
         <?php
-        if (isset($_SESSION['delete-success']) && $_SESSION['delete-success']) {
-            $notification = new Notification('Success!', 'Curriculum was successfully deleted!', 'success');
+        if (isset($_SESSION['notification'])) {
+            $note = $_SESSION['notification'];
+            $notification = new Notification($note['title'], $note['msg'], $note['type']);
             $notification->display();
-            unset($_SESSION['delete-success']);
+            unset($_SESSION['notification']);
         }
         ?>
         <div id="curriculum-btn-group" class="input-group">
-            <a id="new-curriculum-btn" href="/curricula/create">
-                <button class="cpca btn"><i class="fa fa-plus"></i> Create Curriculum</button>
-            </a>
-            <a id="restore-curriculum-btn" class="ml-3" href="/curricula/restore">
-                <button class="btn-outline-secondary btn"><i class="fa fa-repeat"></i> Restore</button>
-            </a>
+            <?php if (hasRole(Role::Coordinator)) { ?>
+                <a id="new-curriculum-btn" href="/curricula/create">
+                    <button class="cpca btn"><i class="fa fa-plus"></i> Create Curriculum</button>
+                </a>
+            <?php
+            }
+            if (hasRole(Role::Admin)) {
+            ?>
+                <a id="restore-curriculum-btn" class="ml-3" href="/curricula/archive">
+                    <button class="btn-outline-secondary btn"><i class="fa fa-archive"></i> See Archive</button>
+                </a>
+            <?php } ?>
         </div><br />
 
         <form id="curriculum-filter" action="/curricula" method="post" class="input-group" style="max-width: 500px; width: 100%; margin: 0 auto">
@@ -68,18 +92,22 @@ if (!empty($params) && $params[0] == 'view') {
                 <div class="card text-center result-card">
                     <div class="card-body">
                         <h4 class="card-title"><?= $r['curriculumname'] ?></h4>
-                        <h6 class="card-subtitle text-muted"><?= $r['curriculumtype'] ?></h6>
                     </div>
                     <div class="card-footer d-flex flex-row justify-content-center">
                         <a href="/curricula/view/<?= $r['curriculumid'] ?>">
                             <button class="btn btn-outline-secondary btn-sm ml-2">View</button>
                         </a>
-                        <a href="/curricula/edit/<?= $r['curriculumid'] ?>">
-                            <button class="btn btn-outline-secondary btn-sm ml-2">Edit</button>
-                        </a>
-                        <a href="/curricula/delete/<?= $r['curriculumid'] ?>">
-                            <button class="btn btn-outline-danger btn-sm ml-2">Delete</button>
-                        </a>
+                        <?php if (hasRole(Role::Coordinator)) { ?>
+                            <a href="/curricula/classes/<?= $r['curriculumid'] ?>">
+                                <button class="btn btn-outline-secondary btn-sm ml-2">Classes</button>
+                            </a>
+                            <a href="/curricula/edit/<?= $r['curriculumid'] ?>">
+                                <button class="btn btn-outline-secondary btn-sm ml-2">Edit</button>
+                            </a>
+                            <a href="/curricula/delete/<?= $r['curriculumid'] ?>">
+                                <button class="btn btn-outline-danger btn-sm ml-2">Delete</button>
+                            </a>
+                        <?php } ?>
                     </div>
                 </div>
                 <?php
