+++ views/curricula/curricula_view.php
@@ -1,5 +1,23 @@
 <?php
+/**
+ * PEP Capping 2017 Algozzine's Class
+ *
+ * Displays details about a curriculum.
+ *
+ * This shows detailed information about the chosen
+ * curriculum. Along with curriculum-related information,
+ * the related classes are displayed as well for
+ * quick access.
+ *
+ * @author Jack Grzechowiak
+ * @copyright 2017 Marist College
+ * @version 0.6
+ * @since 0.1
+ */
+
 global $params, $db;
+
+# Get topic name from params
 $id = $params[1];
 
 $result = $db->query("SELECT * FROM curricula WHERE curriculumid = $1", [$id]);
@@ -13,27 +31,40 @@ if (pg_num_rows($result) == 0) {
 $curricula = pg_fetch_assoc($result);
 pg_free_result($result);
 
-$topics = $db->query("SELECT * FROM curriculumclasses WHERE curriculumid = $1", [$id]);
+$topics = $db->query("SELECT * FROM curriculumclasses, classes WHERE curriculumid = $1 ".
+    "AND curriculumclasses.classid = classes.classid ".
+    "AND classes.df = 0 ORDER BY classes.topicname", [$id]);
+$curriculaName = $curricula['curriculumname'];
+$site = pg_fetch_assoc($db->query("SELECT * FROM sites WHERE sitename = $1", [$curriculaName]));
 
 include('header.php');
 ?>
 <div style="width: 100%">
-    <a href="/back"><button class="cpca btn"><i class="fa fa-arrow-left"></i> Back</button></a>
+    <div class="row">
+			<div class="col">
+				<button class="cpca btn" onclick="goBack()"><i class="fa fa-arrow-left"></i> Back</button>
+			</div>
+			<div class="col pr-5" align="right">
+				<button type="button" class="btn cpca" onclick="window.print()"><i class="fa fa-print" aria-hidden="true"></i> Print</button>
+			</div>
+		</div>
     <div class="form-wrapper card view-card">
         <h4 class="card-header text-left">
             <?= $curricula['curriculumname'] ?>
-            <div class="float-right">
-                <a href="/curricula/edit/<?= $id ?>"><button class="btn btn-outline-secondary btn-sm">Edit</button></a>
-                <a href="/curricula/delete/<?= $id ?>"><button class="btn btn-outline-danger btn-sm">Delete</button></a>
-            </div>
+            <?php if (hasRole(Role::Coordinator)) { ?>
+                <div class="float-right">
+                    <a href="/curricula/edit/<?= $id ?>"><button class="btn btn-outline-secondary btn-sm">Edit</button></a>
+                    <a href="/curricula/delete/<?= $id ?>"><button class="btn btn-outline-danger btn-sm">Delete</button></a>
+                </div>
+            <?php } ?>
         </h4>
         <div class="card-body d-flex justify-content-center flex-column">
             <h4>Information</h4>
             <div class="d-flex justify-content-center">
                 <div class="display-stack">
-                    <div class="display-top"><?= $curricula['curriculumtype'] ?></div>
+                    <div class="display-top"><?= $site['sitetype'] ?></div>
                     <div class="display-split"></div>
-                    <div class="display-bottom">Type</div>
+                    <div class="display-bottom">Location</div>
                 </div>
                 <div class="display-stack">
                     <div class="display-top"><?= $curricula['missnumber'] ?></div>
@@ -51,7 +82,7 @@ include('header.php');
                         <tr>
                             <td class="pl-2 align-middle"><?= $class['topicname'] ?></td>
                             <td class="pr-2 text-right">
-                                <a href="/classes/view/<?= $class['topicname'] ?>">
+                                <a href="/classes/view/<?= $class['classid'] ?>">
                                     <button class="btn btn-outline-secondary btn-sm">View</button>
                                 </a>
                             </td>
