+++ views/locations/locations_view.php
@@ -1,4 +1,20 @@
 <?php
+/**
+ * PEP Capping 2017 Algozzine's Class
+ *
+ * Displays details about a location.
+ *
+ * This page displays the detailed information
+ * about a specific location. From here the user
+ * can also edit or delete the entry.
+ *
+ * @author Jack Grzechowiak
+ * @copyright 2017 Marist College
+ * @version 0.6
+ * @since 0.1
+ * @deprecated
+ */
+
 global $params, $db;
 array_shift($params);
 
@@ -19,14 +35,23 @@ pg_free_result($result);
 include('header.php');
 ?>
     <div style="width: 100%">
-        <a href="/back"><button class="cpca btn"><i class="fa fa-arrow-left"></i> Back</button></a>
+        <div class="row">
+			<div class="col">
+				<button class="cpca btn" onclick="goBack()"><i class="fa fa-arrow-left"></i> Back</button>
+			</div>
+			<div class="col pr-5" align="right">
+				<button type="button" class="btn cpca" onclick="window.print()"><i class="fa fa-print" aria-hidden="true"></i> Print</button>
+			</div>
+		</div>
         <div class="form-wrapper card view-card">
             <h4 class="card-header text-left">
                 <?= $site['sitename'] ?>
-                <div class="float-right">
-                    <a href="/locations/edit/<?= implode('/', $params) ?>"><button class="btn btn-outline-secondary btn-sm">Edit</button></a>
-                    <a href="/locations/delete/<?= implode('/', $params) ?>"><button class="btn btn-outline-danger btn-sm">Delete</button></a>
-                </div>
+                <?php if (hasRole(Role::Coordinator)) { ?>
+                    <div class="float-right">
+                        <a href="/locations/edit/<?= implode('/', $params) ?>"><button class="btn btn-outline-secondary btn-sm">Edit</button></a>
+                        <a href="/locations/delete/<?= implode('/', $params) ?>"><button class="btn btn-outline-danger btn-sm">Delete</button></a>
+                    </div>
+                <?php } ?>
             </h4>
             <div class="card-body d-flex justify-content-center flex-column">
                 <h4>Information</h4>
